Refactor MockConsole out into a console helpers file.
